+++ src/main/java/org/virep/jdabot/commands/game/InventoryCommand.java
@@ -13,7 +13,7 @@ import org.json.JSONTokener;
 import org.virep.jdabot.commands.games.ActivityCommand;
 import org.virep.jdabot.database.Database;
 import org.virep.jdabot.language.Language;
-import org.virep.jdabot.slashcommandhandler.Command;
+import org.virep.jdabot.handlers.Command;
 import org.virep.jdabot.utils.ErrorManager;
 
 import java.awt.*;
