+++ src/main/java/org/virep/jdabot/commands/music/QueueCommand.java
@@ -15,7 +15,7 @@ import org.virep.jdabot.language.Language;
 import org.virep.jdabot.music.AudioManagerController;
 import org.virep.jdabot.music.GuildAudioManager;
 import org.virep.jdabot.music.TrackScheduler;
-import org.virep.jdabot.slashcommandhandler.Command;
+import org.virep.jdabot.handlers.Command;
 
 import java.awt.*;
 import java.time.Instant;
