+++ src/main/java/org/virep/jdabot/commands/music/StopCommand.java
@@ -9,7 +9,7 @@ import net.dv8tion.jda.api.interactions.commands.build.Commands;
 import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
 import org.virep.jdabot.language.Language;
 import org.virep.jdabot.music.AudioManagerController;
-import org.virep.jdabot.slashcommandhandler.Command;
+import org.virep.jdabot.handlers.Command;
 
 import java.util.ArrayList;
 import java.util.Collections;
