+++ src/main/java/org/virep/jdabot/handlers/Command.java
@@ -1,4 +1,4 @@
-package org.virep.jdabot.slashcommandhandler;
+package org.virep.jdabot.handlers;
 
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
