+++ src/main/java/org/virep/jdabot/handlers/SlashHandler.java
@@ -1,4 +1,4 @@
-package org.virep.jdabot.slashcommandhandler;
+package org.virep.jdabot.handlers;
 
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.Guild;
